+++ tests/testthat/helper-capture.R
@@ -1,4 +1,3 @@
-
 named <- function(x) {
   set_names(x, names2(x))
 }